SellApiApp/api-sharing-platform/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-cors-headers==3.14.0
django-filter==23.2.0
pytest==7.3.1
pytest-django==4.5.2
djangorestframework-simplejwt==5.2.2